Fix severe bug when branching with a negative offset
This bug is essentially due to a buffer overflow, in fact we were
shifting a byte to the left by 8 bits.

+++ gork/zoptable.go
@@ -153,7 +153,7 @@ func ZJump(zm *ZMachine, offset uint16) {
 	// this is not a branch instruction
 	// jumping to an instruction in a different routine is permitted,
 	// but the standard consider it bad practice :)
-	zm.seq.pos = zm.CalcJumpAddress(int16(offset))
+	zm.seq.pos = zm.CalcJumpAddress(int32(int16(offset)))
 }
 
 func ZPrint(zm *ZMachine) {
